refactor: ditch `PGame` (#157)
Thanks to the power of category theory, we don't actually have to go through the hassle of PGame in order to define IGame! See the module docstring on the new file CombinatorialGames.Game.Functor for more info.

We'll use this to define loopy games shortly.

Co-authored by: Aaron Liu <aaronliu2008@outlook.com>